Extract `HTTPClientSpy` into a separate file
