+++ Gymondo/GymondoTests/HTTPClientTests.swift
@@ -194,37 +194,6 @@ class APIClientTests: XCTestCase {
         let json = ["results": items]
         return try! JSONSerialization.data(withJSONObject: json)
     }
-
-    class HTTPClientSpy: HTTPClient {
-        private(set) var dispatchCalled = false
-        private(set) var capturedRequests: [URLRequest] = []
-
-        var responseStub: AnyPublisher<Data, NetworkRequestError>?
-
-        func dispatch<ReturnType: Codable>(request: URLRequest) -> AnyPublisher<ReturnType, NetworkRequestError> {
-            dispatchCalled = true
-            capturedRequests.append(request)
-
-            if let responseStub = responseStub {
-                return responseStub
-                    .mapError { _ in
-                        NetworkRequestError.invalidRequest
-                    }
-                    .flatMap { data in
-                        Just(data)
-                            .decode(type: ReturnType.self, decoder: JSONDecoder())
-                            .mapError { error in
-                                NetworkRequestError.decodingError(error.localizedDescription)
-                            }
-                    }
-                    .eraseToAnyPublisher()
-            } else {
-                let errorPublisher = Fail<ReturnType, NetworkRequestError>(error: .error4xx(400))
-                    .eraseToAnyPublisher()
-                return errorPublisher
-            }
-        }
-    }
 }
 
 struct MockEndpointRouter: EndpointRouter {
